Big Legacy Code Overhaul
- Constitutional crises now use factions like they should have all
along. Some of the fancier crisis stuff is in limbo for now.
- Crises are much rarer - democratic rulers can't use them as an
alternative to signing bills, they have to actually dissolve the
legislature to get out of that. Similarly, autocratic rulers can dismiss
MOST demands with no consequences except creating demand for a Senate -
dismissing a demand for a Senate WILL trigger a crisis.
- The existing rule-by-decree decisions have been truncated for now due
to having been discovered to be hugely bugged. They're going to need to
be reworked.
- A bunch of AI changes, mostly to decisions.
- A new, not-yet-fully-implemented faction, which is designed to
discourage hoarding secondary titles (not sure what to do about
destroying secondary titles yet)