Love, Marriage, and Birth Control
- Characters will now generally try to marry for love. A "Not Ready to
Marry" trait prevents them from immediately engaging in political
marriages, but this trait goes away with time, and in the future it will
be possible to try to convince your children to agree to an arranged
marriage
- Population explosion had become a major problem. I spent two days
trying various ways to fix the issue before finally running a dead
chicken over the code, and now population explosion is no longer a huge
issue and the game motors along at least as far as 3100. We may never
know why.
- I partially rearranged the council jobs so that the AI will actually
do things. It's sending councillors on jobs they can't actually do now,
but it's still an improvement.
- I made a vain attempt to force republic characters to change their
clothes based on circumstance, but this appears to be a lost cause short
of Paradoxian intervention.