Bianjiang and Mercenaries
- Added the Republic of Bianjiang. Whereas Xin Zizhiqu was open to
public settlement, the settlement of Bianjiang was conducted by the
Xingjie government. The last colonial governor, Yong Jia, was terrified
of the possibility that the Bianjiangese might join the Colonial
Separatist movement, and as such declared martial law in the years just
prior to the Crisis. Unfortunately for Jia, many of the military
governors he appointed had Separatist sympathies themselves. Admiral
Qian Deng led a military coup, imprisoning Jia and establishing a
military junta that now governs Bianjiang and is a close ally to the
Orion Reach.
- Added a number of private military companies (mercenaries) scattered
about the galaxy. I've tried to give the various PMCs diverse
"personalities" that define what sort of contracts they are willing to
accept and what mix of units they field. For example, Aegis Security
will not support aggressive wars, while Zahran Defense Industries has a
grudge against Tierra Celeste and will not support Celestean rules.
- Related to the above: It is now possible for the leaders of orphaned
fleets or navies (that is, the militaries of states that no longer
exist) to elect to become mercenaries. The specific inspiration here is
the Eridani Light Horse of the Battletech universe.
- Made the AI create kingdom titles as emperors. This results in the
Republic of the Centauri Reach being created, and it hasn't done
anything about emperors HOARDING kingdom titles, but at least now they
won't DESTROY them, so we can hypothetically organize the AI to try to
get them back.
- Fixed numerous bugs with the character history files.
- Made yet more changes to the cloning AI. The AI will no longer queue
up multiple clones at once, which was presumably the cause of the "I
MUST HAVE ONE HUNDRED THOUSAND CLONES" bug. Just in case, I've also set
a hard cap - the AI won't clone itself once it already has five
children.
- Restored the old general opinion bonuses/maluses for moral authority.
This isn't perfect - I'd like if it only applied to followers of other
ideologies, and also were applied equally to rulers and non-rulers - but
it does give more consequence to low moral authority. As it was, people
seemed totally fine with Morgan losing the entire periphery.
- Buffed Juan Lebedev's health substantially. He's an old man, and
Matias is apparently incapable of winning the revolution, so Lebedev
would frequently die of natural causes and end the big civil war in an
anticlimactic Confederate victory. Something similar may be necessary
for Ryusei Sato, who has a similar habit of croaking early.
- Tried an experiment with republican clothing. It does not appear to
have worked, sadly.